requirements.txt
alembic==1.14.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
openai==1.54.3
SQLAlchemy==2.0.36
pyTelegramBotAPI==4.24.0
boto3==1.35.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
